Added the SLDEditor widget class wrapping the Qgraphicsview and the future toolbar
