+++ examples/quickrefl/core/editors/sldeditor/CMakeLists.txt
@@ -21,6 +21,8 @@ target_sources(${library_name} PRIVATE
     graphicsscene.h
     sldviewwidget.h
     sldviewwidget.cpp
+    sldeditor.h
+    sldeditor.cpp
 )
 
 target_include_directories(${library_name} PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}>)
