+++ examples/quickrefl/core/mainwindow/reflwidgets.cpp
@@ -8,7 +8,7 @@
 // ************************************************************************** //
 
 #include "reflwidgets.h"
-#include "sldviewwidget.h"
+#include "sldeditor.h"
 #include "styleutils.h"
 #include <QVBoxLayout>
 
@@ -17,26 +17,6 @@
 LayerCanvas::LayerCanvas(QWidget* parent) : QWidget(parent)
 {
     setWindowTitle(QString("Layer Canvas"));
-}
-
-//! ---------------------------------------------------------------------------
-
-SLDEditor::SLDEditor(SLDViewWidget* sld_view, QWidget* parent) : QWidget(parent)
-{
-    setWindowTitle(QString("SLD Editor"));
-    auto layout = new QVBoxLayout;
-    layout->addWidget(sld_view);
-    setLayout(layout);
-}
-
-QSize SLDEditor::sizeHint() const
-{
-    return StyleUtils::DockSizeHint();
-}
-
-QSize SLDEditor::minimumSizeHint() const
-{
-    return StyleUtils::DockMinimumSizeHint();
 }
 
 //! ---------------------------------------------------------------------------
